:art: Enriched EasingHelper.

+++ API/src/main/java/xyz/bitsquidd/bits/lib/helper/math/EasingHelper.java
@@ -3,7 +3,7 @@ package xyz.bitsquidd.bits.lib.helper.math;
 import org.jetbrains.annotations.Range;
 
 public final class EasingHelper {
-    private static final double EASING_MAGIC = 1.70158;
+    private static final float EASING_MAGIC = 1.70158f;
 
     private static final double PI = Math.PI;
     private static final double HALF_PI = PI / 2;
@@ -18,7 +18,7 @@ public final class EasingHelper {
      *
      * @return the eased value
      */
-    public static double easeLinear(@Range(from = 0, to = 1) double t) {
+    public static float easeLinear(@Range(from = 0, to = 1) float t) {
         return t;
     }
 
@@ -33,25 +33,25 @@ public final class EasingHelper {
      *
      * @return the eased value
      */
-    public static double easeIn(@Range(from = 0, to = 1) double t, int index) {
-        return Math.pow(t, index);
+    public static float easeIn(@Range(from = 0, to = 1) float t, float index) {
+        return (float)(Math.pow(t, index));
     }
-
+    
     /**
      * Ease out with a custom power
      */
-    public static double easeOut(@Range(from = 0, to = 1) double t, int index) {
-        return 1 - Math.pow(1 - t, index);
+    public static float easeOut(@Range(from = 0, to = 1) float t, int index) {
+        return (float)(1 - Math.pow(1 - t, index));
     }
 
     /**
      * Ease in-out with a custom power
      */
-    public static double easeInOut(@Range(from = 0, to = 1) double t, int index) {
+    public static float easeInOut(@Range(from = 0, to = 1) float t, float index) {
         if (t < 0.5) {
-            return Math.pow(2 * t, index) / 2;
+            return (float)(Math.pow(2 * t, index) / 2f);
         } else {
-            return 1 - Math.pow(-2 * t + 2, index) / 2;
+            return (float)(1 - Math.pow(-2 * t + 2, index) / 2f);
         }
     }
 
@@ -64,22 +64,22 @@ public final class EasingHelper {
      *
      * @return the eased value
      */
-    public static double easeInSine(@Range(from = 0, to = 1) double t) {
-        return 1 - Math.cos((t * PI) / 2);
+    public static float easeInSine(@Range(from = 0, to = 1) float t) {
+        return (float)(1 - Math.cos((t * PI) / 2f));
     }
 
     /**
      * Ease out sine function.
      */
-    public static double easeOutSine(@Range(from = 0, to = 1) double t) {
-        return Math.sin((t * PI) / 2);
+    public static float easeOutSine(@Range(from = 0, to = 1) float t) {
+        return (float)(Math.sin((t * PI) / 2f));
     }
 
     /**
      * Ease in-out sine function.
      */
-    public static double easeInOutSine(@Range(from = 0, to = 1) double t) {
-        return -(Math.cos(PI * t) - 1) / 2;
+    public static float easeInOutSine(@Range(from = 0, to = 1) float t) {
+        return (float)(-(Math.cos(PI * t) - 1) / 2f);
     }
 
 
@@ -92,25 +92,25 @@ public final class EasingHelper {
      *
      * @return the eased value
      */
-    public static double easeInExpo(@Range(from = 0, to = 1) double t) {
-        return t == 0 ? 0 : Math.pow(2, 10 * (t - 1));
+    public static float easeInExpo(@Range(from = 0, to = 1) float t) {
+        return (float)(t == 0 ? 0f : Math.pow(2, 10 * (t - 1)));
     }
 
     /**
      * Ease out exponential function.
      */
-    public static double easeOutExpo(@Range(from = 0, to = 1) double t) {
-        return t == 1 ? 1 : 1 - Math.pow(2, -10 * t);
+    public static float easeOutExpo(@Range(from = 0, to = 1) float t) {
+        return (float)(t == 1 ? 1f : 1 - Math.pow(2, -10 * t));
     }
 
     /**
      * Ease in-out exponential function.
      */
-    public static double easeInOutExpo(@Range(from = 0, to = 1) double t) {
+    public static float easeInOutExpo(@Range(from = 0, to = 1) float t) {
         if (t == 0) return 0;
         if (t == 1) return 1;
-        if (t < 0.5) return Math.pow(2, 20 * t - 10) / 2;
-        return (2 - Math.pow(2, -20 * t + 10)) / 2;
+        if (t < 0.5) return (float)(Math.pow(2, 20 * t - 10) / 2);
+        return (float)((2 - Math.pow(2, -20 * t + 10)) / 2f);
     }
 
 
@@ -123,17 +123,17 @@ public final class EasingHelper {
      *
      * @return the eased value
      */
-    public static double easeInBack(@Range(from = 0, to = 1) double t) {
+    public static float easeInBack(@Range(from = 0, to = 1) float t) {
         double c3 = EASING_MAGIC + 1;
-        return c3 * t * t * t - EASING_MAGIC * t * t;
+        return (float)(c3 * t * t * t - EASING_MAGIC * t * t);
     }
 
     /**
      * Ease out back function. Creates an overshooting effect at the end.
      */
-    public static double easeOutBack(@Range(from = 0, to = 1) double t) {
+    public static float easeOutBack(@Range(from = 0, to = 1) float t) {
         double c3 = EASING_MAGIC + 1;
-        return 1 + c3 * Math.pow(t - 1, 3) + EASING_MAGIC * Math.pow(t - 1, 2);
+        return (float)(1 + c3 * Math.pow(t - 1, 3) + EASING_MAGIC * Math.pow(t - 1, 2));
     }
